translation-validator/requirements.txt
google-adk
google-cloud-aiplatform
google-cloud-storage
google-cloud-bigquery
python-dotenv
google-cloud-aiplatform[adk,agent_engines]